Rename type for clarity, move to more appropriate package
